+++ tasky/package.json
@@ -27,10 +27,8 @@
     "@emotion/styled": "^11.10.6",
     "@mui/icons-material": "^5.11.11",
     "@mui/material": "^5.11.13",
-    "dotenv": "^16.0.3",
-    "firebase": "^9.19.1",
-    "react-form-stepper": "^2.0.3",
     "babel-plugin-styled-components": "^2.1.1",
+    "firebase": "^9.19.1",
     "styled-react-modal": "^3.0.1"
   }
 }
