+++ tasky/src/pages/index.js
@@ -1,5 +1,7 @@
 import Auth from "./auth"
 
+
+
 export default function Home() {
   return Auth();
 }
